fixed pyproject.toml and simplified experiment_run dir structure

+++ src/ada_verona/robustness_experiment_box/database/verification_context.py
@@ -78,9 +78,13 @@ class VerificationContext:
         """
         Delete the temporary path and its contents.
         """
-      
-
-        self.tmp_path.unlink()
+        import shutil
+        
+        if self.tmp_path.exists():
+            if self.tmp_path.is_dir():
+                shutil.rmtree(self.tmp_path)
+            else:
+                self.tmp_path.unlink()
 
 
     def save_status_list(self, epsilon_status_list: list[EpsilonStatus]) -> None:
